Menu: extract the Files tab into build_files.cpp
Verbatim move of the GIF slot-management rows and the landing-page
background library (import/apply/delete + preview panel). The slot rows
come from ctx.gif_slot_row; build_menu keeps constructing the shared
GIF preview machinery.
